refactor(bal-time-input): moved time format util to bal-time-input dir
